Updated storage methods, added optional access lists
#User:
Access can now be obtained/deleted through /access

##GET /as/access
Create a new User with a upload path. User-UUID and Uploadpath-UUID are
not the same

##DELETE /as/access/{User-UUID}
Deletes User and all his files

#File Up-/Download & Delete

##POST /as/{Uploadpath-UUID}/USERSPECIFIED/{User-UUID}
Uploads a file via multipart/form-data, the key "file" is for the
filename
User must have the right to upload to the uploadpath, defaults to the
owner

##GET /as/{Uploadpath-UUID}/USERSPECIFIED
Downloads an uploaded file
if no http access list is specified for the Uploadpath, everyone can
download the file, if an acl is defined, only those specified can
download

##DELETE /as/{Uploadpath-UUID}/USERSPECIFIED/{User-UUID}
Deletes the given file if the User has the right to delete it.